Add files via upload
Hi guys, this is the draft for our project part 2 user interface mockups. It is still quite rough at this point, but it contains the structure and things that are needed for the user interface. Feel free to add or modify it and talk about your thoughts on discord. I think it will be a good idea that we do the Storyboard Sequences after we did the user interface mockups cause we can then label them and make it easier. Also, please get things done before Thursday so that we will have a chance to meet the TA and ask questions about the upcoming due.